Issue 1120 fixed. Test cases updated. Issue 1093 fixed. UI Issue.
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@11243 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/test/org/mifos/application/accounts/loan/business/TestLoanBO.java
@@ -4093,7 +4093,7 @@ public class TestLoanBO extends MifosTestCase {
 		loan = (LoanBO) accountBO;
 		assertEquals(Integer.valueOf("0"), loan.getPerformanceHistory()
 				.getTotalNoOfMissedPayments());
-		assertEquals(Integer.valueOf("0"), loan.getPerformanceHistory()
+		assertEquals(Short.valueOf("0"), loan.getPerformanceHistory()
 				.getDaysInArrears());
 		assertEquals(Integer.valueOf("0"), loan.getPerformanceHistory()
 				.getNoOfPayments());
